+++ crates/amaru/src/bin/amaru/cmd/snapshot/create/mod.rs
@@ -33,6 +33,7 @@ use amaru_mithril::{
     parse_header_slot_and_hash,
 };
 use amaru_observability::info;
+use amaru_progress_bar::TerminalProgressBar;
 use anyhow::anyhow;
 use clap::{ArgAction, Parser};
 use serde::{Deserialize, Serialize};
@@ -316,7 +317,10 @@ async fn run(args: Args) -> Result<(), Box<dyn std::error::Error>> {
             );
         } else {
             info!(cli::mithril::DOWNLOAD, from_chunk, target_dir = %cardano_node_db.display());
-            download_from_mithril(network, cardano_node_db.clone(), from_chunk).await?;
+            download_from_mithril(network, cardano_node_db.clone(), from_chunk, |length, template| {
+                TerminalProgressBar::new(length as u64, template).boxed()
+            })
+            .await?;
         }
 
         Some(SnapshotBuildContext {